Magnetization update allowing change in magnetic field direction.
Updated magnetize and magnetize2 programs to accept the direct of the magnetic field using spherical coordinates: theta and phi.
B_min, B_max, B_inc, and B_rate now all refer to the magnitude of the magnetic field, B.
B_theta and B_phi are given in degrees (and the code converts to radians).