MSD Server now supports /reset request
Updated JS, Java, and Python to handle a reinitialize or randomize with a potential reseed, or custom seed.

Also added the ability to get the seed with the state, and added support to MSDWorker.py and MSDWorker.java for only returning the partial state. This may be leveraged in the future to allow for faster HTTP communication and reduced overhead. No API hook has been created yet, however.