Responsive Three.js canvas
The Three.js canvas now auto sizes to fill the container. Uses window resize event to actually modify the underlying size of the canvas so it can be rendered in the new native resolution. Can add maxCanvasWidth and maxCanvasHeight to clamp the resolution, but this is off by default.
Also added black bars (background-color) to fill out the horizontal space.